requirements.txt
func_model.egg==info
nibabel==5.1.0
pandas==2.1.3
setuptools==65.6.3
